fastapi/requirements.txt
fastapi==0.111.0
matplotlib==3.8.3
numpy==1.26.4
pillow==10.2.0
python-multipart==0.0.9
tensorflow==2.16.1
tensorflow-serving-api==2.16.1
uvicorn==0.30.1